refactor: 장바구니 TableViewCell 클래스 분리
